add some translations to points charts
