+++ frontend/components/Dashboard/PointsListItemCard.tsx
@@ -1,4 +1,4 @@
-import { useState } from "react"
+import { useCallback, useMemo, useState } from "react"
 
 import { Grid } from "@mui/material"
 import { styled } from "@mui/material/styles"
@@ -7,10 +7,13 @@ import PointsItemTable from "./PointsItemTable"
 import { FormSubmitButton } from "/components/Buttons/FormSubmitButton"
 import PointsProgress from "/components/Dashboard/PointsProgress"
 import { CardSubtitle, CardTitle } from "/components/Text/headers"
+import ProfileTranslations from "/translations/profile"
 import formatPointsData from "/util/formatPointsData"
+import { useTranslator } from "/util/useTranslator"
 
 import {
   CourseCoreFieldsFragment,
+  UserCoreFieldsFragment,
   UserCourseProgressCoreFieldsFragment,
   UserCourseServiceProgressCoreFieldsFragment,
 } from "/graphql/generated"
@@ -21,12 +24,6 @@ const PointsListItemCardContainer = styled(Grid)`
   padding: 1rem;
 `
 
-interface PersonalDetails {
-  firstName: string
-  lastName: string
-  email: string
-  sid: string
-}
 interface PointsListItemCardProps {
   course?: CourseCoreFieldsFragment | null
   userCourseProgress?: UserCourseProgressCoreFieldsFragment | null
@@ -35,48 +32,63 @@ interface PointsListItemCardProps {
     | null
   cutterValue?: number
   showPersonalDetails?: boolean
-  personalDetails?: PersonalDetails
+  user?: UserCoreFieldsFragment
   showProgress?: boolean
 }
 
 interface PersonalDetailsDisplayProps {
-  personalDetails: PersonalDetails
+  user: UserCoreFieldsFragment
 }
 
-const PersonalDetailsDisplay = (props: PersonalDetailsDisplayProps) => {
-  const { personalDetails } = props
+const PersonalDetailsDisplay = ({ user }: PersonalDetailsDisplayProps) => {
+  const t = useTranslator(ProfileTranslations)
+
   return (
     <>
       <CardSubtitle>
-        Name: {personalDetails.firstName} {personalDetails.lastName}
+        {t("name")} {user.full_name}
       </CardSubtitle>
-      <CardSubtitle>e-mail: {personalDetails.email}</CardSubtitle>
-      <CardSubtitle>student number: {personalDetails.sid}</CardSubtitle>
+      <CardSubtitle>
+        {t("email")} {user.email ?? "-"}
+      </CardSubtitle>
+      <CardSubtitle>
+        {t("studentNumber")} {user.real_student_number ?? "-"}
+      </CardSubtitle>
     </>
   )
 }
 function PointsListItemCard(props: PointsListItemCardProps) {
+  const t = useTranslator(ProfileTranslations)
   const {
     course,
     userCourseProgress,
     userCourseServiceProgresses,
     cutterValue = 0,
     showPersonalDetails,
-    personalDetails,
+    user,
     showProgress = true,
   } = props
   const [showDetails, setShowDetails] = useState(false)
 
   // TODO: do this in the backend
-  const formattedPointsData = formatPointsData({
-    userCourseProgress,
-    userCourseServiceProgresses,
-  })
+  const formattedPointsData = useMemo(
+    () =>
+      formatPointsData({
+        userCourseProgress,
+        userCourseServiceProgresses,
+      }),
+    [userCourseProgress, userCourseServiceProgresses],
+  )
+
+  const onShowDetailsClick = useCallback(
+    () => setShowDetails((prev) => !prev),
+    [setShowDetails],
+  )
 
   return (
-    <PointsListItemCardContainer item sm={12} lg={12}>
-      {showPersonalDetails && personalDetails ? (
-        <PersonalDetailsDisplay personalDetails={personalDetails} />
+    <PointsListItemCardContainer item xs={12} sm={12} md={12} lg={12}>
+      {showPersonalDetails && user ? (
+        <PersonalDetailsDisplay user={user} />
       ) : (
         <CardTitle component="h2" variant="h3">
           {course?.name}
@@ -86,16 +98,15 @@ function PointsListItemCard(props: PointsListItemCardProps) {
         <>
           <PointsProgress
             percentage={formattedPointsData.total * 100}
-            title="Total progress"
+            title={t("totalProgress")}
           />
           <PointsProgress
             percentage={formattedPointsData.exercises * 100}
-            title="Exercises completed"
+            title={t("exercisesCompleted")}
           />
-          <hr />
         </>
       )}
-      {Object.keys(formattedPointsData?.groups ?? {}).length ? (
+      {Object.keys(formattedPointsData?.groups ?? {}).length > 0 ? (
         <>
           <PointsItemTable
             studentPoints={formattedPointsData.groups}
@@ -104,14 +115,14 @@ function PointsListItemCard(props: PointsListItemCardProps) {
           />
           <FormSubmitButton
             variant="text"
-            onClick={() => setShowDetails(!showDetails)}
+            onClick={onShowDetailsClick}
             fullWidth
           >
-            {showDetails ? "show less" : "show detailed breakdown"}
+            {showDetails ? t("showLess") : t("showDetailedBreakdown")}
           </FormSubmitButton>
         </>
       ) : (
-        <p>No points data available</p>
+        <p>{t("noPointsData")}</p>
       )}
     </PointsListItemCardContainer>
   )
