+++ frontend/components/User/Points/PointsListGrid.tsx
@@ -1,3 +1,5 @@
+import { useMemo } from "react"
+
 import Grid from "@mui/material/Grid"
 
 import PointsListItemCard from "/components/Dashboard/PointsListItemCard"
@@ -12,15 +14,19 @@ interface GridProps {
 function PointsListGrid(props: GridProps) {
   const { data, showOnlyTen } = props
 
-  const progressesToShow = showOnlyTen
-    ? (data.currentUser?.progresses ?? []).slice(0, 10)
-    : data.currentUser?.progresses ?? []
+  const progressesToShow = useMemo(
+    () =>
+      showOnlyTen
+        ? (data.currentUser?.progresses ?? []).slice(0, 10)
+        : data.currentUser?.progresses ?? [],
+    [data, showOnlyTen],
+  )
 
   return (
     <Grid container spacing={3}>
-      {progressesToShow.map((progress, index) => (
+      {progressesToShow.map((progress) => (
         <PointsListItemCard
-          key={`${progress.course?.id}-${index}`}
+          key={`${progress.course?.id}-${progress?.user_course_progress?.id}`}
           course={progress.course}
           userCourseProgress={progress.user_course_progress}
           userCourseServiceProgresses={progress.user_course_service_progresses}
